Removed interface HierPermMappingDAO
